refactor(slice6b): als-tempo-automation nach master-timeline/ (Folder-Size, regressions-neutral, pure rename)
