+++ src/automation/master-timeline/als-tempo-automation.ts
